Move CONSTRUCT schema computation to package object
